Fixed the last remainder of int->long switch in the Chapter model

+++ Ogma3/Data/Models/Chapter.cs
@@ -38,6 +38,6 @@ namespace Ogma3.Data.Models
         public  CommentsThread CommentsThread { get; set; }
         
         [Required]
-        public int StoryId { get; set; }
+        public long StoryId { get; set; }
     }
 }
